pyproject.toml
[project]
name = "service_npa"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "theycallmemax", email = "-"},
]
dependencies = [
    "fastapi==0.112.1",
    "uvicorn==0.30.6",
    "pymilvus==2.4.6",
    "langchain==0.2.16",
    "langchain-community==0.2.16",
    "langchain-core==0.2.38",
    "langchain-text-splitters==0.2.4",
    "tqdm==4.66.2",
    "torch==2.4.0",
    "transformers==4.44.0",
    "requests==2.32.3",
    "llamaapi==0.1.36",
    "json5==0.9.17",
    "python-dotenv==1.0.1",
    "sentence-transformers==3.0.1",
    "tf_keras==2.17.0",
    "streamlit==1.38.0",
    
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
